feat. 최근 검색 기록 뷰로 분리
