+++ Particle/Particle/Sources/Main/SearchTab/SearchRouter.swift
@@ -13,37 +13,17 @@ protocol SearchInteractable: Interactable {
 }
 
 protocol SearchViewControllable: ViewControllable {
-    func addSearchResult(_ view: UIViewController)
-    func removeSearchResult()
+    func showSearchResult()
+    func hiddenSearchResult()
 }
 
 final class SearchRouter: ViewableRouter<SearchInteractable, SearchViewControllable>, SearchRouting {
     
-    private let searchResultBuilder: SearchResultBuildable
-    private var searchResultRouter: SearchResultRouting?
-    
     init(
         interactor: SearchInteractable,
-        viewController: SearchViewControllable,
-        searchResultBuildable: SearchResultBuildable
+        viewController: SearchViewControllable
     ) {
-        self.searchResultBuilder = searchResultBuildable
         super.init(interactor: interactor, viewController: viewController)
         interactor.router = self
-    }
-    
-    func attachSearchResult() {
-        guard searchResultRouter == nil else { return }
-        let router = searchResultBuilder.build(withListener: interactor)
-        viewController.addSearchResult(router.viewControllable.uiviewController)
-        attachChild(router)
-        searchResultRouter = router
-    }
-    
-    func detachSearchResult() {
-        guard let searchResult = searchResultRouter else { return }
-        viewController.removeSearchResult()
-        detachChild(searchResult)
-        searchResultRouter = nil 
     }
 }
